pynacl/requirements.txt
cffi==1.12.2
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
